Make Home component load the info it needs
Instead of making other components set the state for them, navegable
components (ie, the ones that are root components in the Router) must
load the info they need upon mount, since they may be loaded in a
fresh page load without a previous state.

We should extend this model to every other page on the app.