Skip programs index if a single one is available
We previously were making an action - that can be called from
different pages of the app - decide to navigate when there was
a single program for the user. This was incorrect - if a user
loads `/programs/:code/deployments`, we need to load the programs
but we don't have to make them go to the programs index.

We've now moved the "skip the index if there's a single option"
logic to the index itself.