Move lambdas dependencies to a layer
Instead of zipping all the dependencies with each of the lambdas, we
put the commmon dependencies on a Layer that the functions depend on.

This saves bandwith - we avoid duplicate uploads - and it allows us to
use AWS Console's online editor to quickly iterate functions online.

We use a temporary `python` directory because Python's path on AWS Lambda
includes `/opt/python`, and the layer itself is extracted in `/opt`.

See #95