Support containers in multiple networks from proxy
The previous script to create the nginx resolver file concatenated
every IP a container had togheter, so when the proxy tried to
forward requests it got something like `172.25.1.1010.15.12.6` as
a hostname - which didn't resolve.

We now only care about the project-specific network created by
docker-compose.

We've also changed how to query which containers do we need to resolve.