Correctly copy Python dependencies to the lambda containers
`docker cp`'s behaviour with directories changes depending on
whether the target directory exists or not.

We need dependencies to get copied to the `package` directory
(without a `site-packages` subdirectory), so now the `docker
cp` works as it should regardless of the `./lambda/package`
being there or not.

This could be further improved by keeping dependencies in a
volume that never gets to the host's filesystem, but that
would also require changing how we zip and upload lambdas'
dependencies layer.