Work around bug in content editor.
A bug in the content editor can cause edits to be unexpectedly lost, due
to a bug in the "Save" / "Discard" handling.

Changes should accumulate until the user clicks "Save", which of course
would commit the changes, or "Discard" which throws them away. The bug
is that when the user clicks the trash can next a message, the message
is _immediately_ delted, and the content is refreshed from the server.
Any changes other than that deletion are simply lost.

The proper fix would be to accumulate the changes, and commit them all
at once. As the "Save" / "Discard" implicitly promise. Unfortunately,
the bug is very tighly integrated into the code, including the
server-side code, and the level of reverse engineering and proper design
needed to fix this is a lot, and could delay protecting users from this
bug.

As a workaround, this change disables the "Delete Message" button when
there are other outstanding changes. And it disables the "Delete
Playlist" button when there is only one playlist left. It changes the
buttons (which are trash can icons) to grey, and adds an informative
tooltip to help the user understand what's going on.

This change improves testing by removing the invalid ProgramId "My test
program 8" and replacing it with "MY-TEST-8".