Fix for an apparent Docker change.
It seems that 'docker-compose build' changed the way that it creates
default container names, in a way that broke the dev-setup.sh script.
The change was to use '-' instead of '_' to concatenate the directory
name and service name. This script relied on that '_', which was nice
defensive programming, and Docker broke it.