Update the Lambda build process to avoid Docker.
The Docker command fails in some environments. Because it was only used
to get a list of the Lambda functions, it was not really necessary, so
it has been replaced with a list of Lambda functions.

In the future this whole mess should go away, to be replaced by a more
streamlined set of Lambda functions, with independent updates for each
function.