Make the program selector more friendly.
Previously, programs were selected by program-id. But that is not something
that users need to know or care about. Even Amplio staff, looking at all
programs, tend to think of them by their human-readable name.

With this change the grid of identical icons now shows the program name, not
the program-id. Optionally (for support folks), the program-id can also be
shown.

In the drop-down menu for selecting programs, only the friendly name is shown.
Because some of them are long, we need to widen the drop-down widget when
it opens, and then narrow it when it closes.